Move mock response to stubs within tests folder
